Fix fatal parse error.
Fix an issue that was occurring when jsdoc echoed output that was not valid JSON. This would cause a fatal error and halt the program, even though the standard approach would be to return the error with the callback. 

Invalid output might occur if a file is not documented or if there is an issue opening the file.